+++ navigation/AuthStack.js
@@ -1,10 +1,10 @@
 import React, { useState, useEffect, useContext } from 'react'
 import AsyncStorage from '@react-native-async-storage/async-storage'
 import { createStackNavigator } from '@react-navigation/stack'
-import AuthLoadingScreen from 'screens/session/AuthLoadingScreen'
-import AuthScreen from 'screens/session/AuthScreen'
-import SignUpScreens from 'screens/session/SignUpScreens'
-import SignInScreen from 'screens/session/SignInScreen'
+import AuthLoadingScreen from '@screens/session/AuthLoadingScreen'
+import AuthScreen from '@screens/session/AuthScreen'
+import SignUpScreens from '@screens/session/SignUpScreens'
+import SignInScreen from '@screens/session/SignInScreen'
 import AppNavigator from './AppNavigator'
 import Context from '@context/context'
