+++ screens/InitAppScreen.js
@@ -4,7 +4,7 @@ import { Avatar } from 'react-native-elements'
 import Proptypes from 'prop-types'
 import Context from '@context/context'
 import AsyncStorage from '@react-native-async-storage/async-storage'
-import getCurrentLocation from '../functions/getCurrentLocation'
+import getCurrentLocation from '@functions/getCurrentLocation'
 
 const InitAppScreen = ({ navigation }) => {
   const context = useContext(Context)
